ml/requirements.txt
# Python 3.12 Dependencies
# Safe Street ML - SOS Detection System

# Core Computer Vision & AI Libraries
tensorflow==2.17.1
keras==3.8.0
numpy==1.26.4
opencv-python==4.10.0.84
mediapipe==0.10.14
scikit-learn==1.6.1
cvlib==0.2.7

# Data Processing & Visualization
pandas==2.3.1
matplotlib==3.9.4
pillow==11.3.0

# Voice & Audio Processing
SpeechRecognition==3.14.3
pyaudio==0.2.14
fuzzywuzzy==0.18.0
python-Levenshtein==0.27.3

# Flask Web Server
flask==3.1.1
requests==2.32.4
python-dotenv==1.0.1

# Database
pymongo==4.13.2

# Hardware Communication
pyserial==3.5